requirements.txt
Django==5.1.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
psycopg2-binary==2.9.10
python-decouple==3.8
Pillow==11.1.0
django-cors-headers==4.4.0
